Fixed animations, then in 15 mins, revamped littereally the entire object system.
I have no idea what happened, even 30 mins after doing it but it works.

Pretty much instead of having a bunch of interfaces that an object is
part of, now an object inherits a single abstract class it is part of,
like Animation or Collider.  To note: Dynamic objects are still the
same.  Yeah.  this is... fun.

TODO: fix everything
TODO: make a collider animation class and an object animation class.